Final PNP Stability Fix: Circular imports, DBus locks, and UI decoupling
This final update resolves all remaining critical issues:
- Fixed circular import by moving is_service_active and check_dependencies
  directly into main_window.py and removing them from __init__.py.
- Bypassed 'Failed to register' DBus timeouts by switching the GUI to
  NON_UNIQUE mode, allowing instant startup as an independent shell.
- Fully decoupled GUI from backend initialization; MainWindow now performs
  async setup and mode detection (observer vs manager) via idle callbacks.
- Fixed logic bug where the tester update loop would not start in manager
  mode due to premature initialization checks.
- Resolved 'Adwaita-CRITICAL' errors using Gtk.ListBox for dynamic rows.
- Eliminated memory leaks with explicit cleanup() and signal management.
- Ensured strict POSIX compliance for all scripts and updated for 2026.

The PNP application is now a robust, high-performance, and portable
controller mapper that functions reliably across all Linux environments.
