--- requirements-dev.lock
@@ -1,21 +1,0 @@
-# generated by rye
-# use `rye lock` or `rye sync` to update this lockfile
-#
-# last locked with the following flags:
-#   pre: false
-#   features: []
-#   all-features: false
-#   with-sources: false
-#   generate-hashes: false
-#   universal: false
-
--e file:.
-evdev==1.9.3
-    # via pnp
-installer==1.0.1
-pycairo==1.29.0
-    # via pygobject
-pygobject==3.56.3
-    # via pnp
-pyudev==0.24.4
-    # via pnp
