+++ src/pnp/main.py
@@ -1,4 +1,3 @@
-import os
 import sys
 import argparse
 import logging
